Add logging and error handling to Kafka stream creation
This commit adds logging to the process of Kafka stream creation in the SensorStreamManager service, to better track the system's activities. It also includes error handling for the case where the specified Kafka topic or partition is unknown. Furthermore, this update enhances exception handling and provides more effective feedback interface for the user with the use of printlines to track system processes. Lastly, the scalatest version was updated for better testing capabilities.